Re-factors: fetch colours, skip hex strings
Utility functions have been added to handle the fetching of colour hex
strings from a theme.

Colour values can now be specified as hex strings (e.g. "#123456"). If
they are names then they will be converted to the corresponding hex
string from the theme. If they are hex strings already they will be
returned as-is.