Adds =completing-read=, global mapping variable
=completing-read= is now used for both arguments to
=modus-utils-insert-theme-colours-at-point= so that the current Emacs
completion framework can be used to aid selection of the correct
symbols. Thanks to Protesilaos Stavrou <public@protesilaos.com> for this
patch.

Also adds a new variable =modus-utils-export-functions=. This variable
replaces the local mapping variable in =modus-utils-export-theme=
allowing users to define this mapping themselves. It also provides a
source for the above completing-read for the export format.