Renames modus-utils to modus-exporter
The scope of this library is simply to allow modus-themes to be
exported. While these functions can be defined as utilities, other types
of utilities relating to these themes could also be created. The library
name has therefore been changed to reflect this.

This library could be included as part of a larger collection of
utilities which itself could be named modus-utils.