Starts conversion to a multi-file package
modus-exporter is now a multi-file package.

The modus-exporter.el file in the base directory contains the main
functionality, but the Alacritty export function has been moved to a
separate file, ./exporters/alacritty.el. When modus-exporter.el is
loaded, all .el* files in ./exporters are loaded automatically.

Exporter files should define their specific export function and also add
the function to the =modus-exporter-export-functions= list if it does
not already exist.

modus-exporter.el now contains a new constant =modus-exporter-base=
which stores the absolute file-system directory of the package. This is
used for locating the exporter files. This constant is initialised
either with =load-file-name= for cases where the package is being loaded
or =buffer-file-name= for cases where the buffer is being evaluated.
This allows the package either to be loaded using regular methods of
loading packages or to be evaluated directly from the loaded buffer.

Ref:
https://www.gnu.org/software/emacs/manual/html_node/elisp/Multi_002dfile-Packages.html#Multi_002dfile-Packages